add label field/column for seasons
The frontend needs to know what each
season represents. I didn't want to use the
'name' field since it's language specific.
The previous solution was to use the ids, but
it required hardcoding the ids in frontend
(id 1 == spring, etc).
Labels are not language specific, so we can use them
instead.